2.90 fixed discounts.Test needed||

+++ Documents/NetBeansProjects/DynaShoppingCart_2.90/src/main/java/com/dynamotors/dynashoppingcart/entities/OrderItems.java
@@ -14,8 +14,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.NamedQueries;
-import javax.persistence.NamedQuery;
 import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
@@ -48,7 +46,7 @@ public class OrderItems implements Serializable {
     @Column(name = "item_price_retail")
     private double itemPriceRetail;     
     @Column(name = "item_discount")
-    private int itemDiscount;     
+    private double itemDiscount;     
     @Size(max = 100)
     @Column(name = "item_size", length = 100)
     private String itemSize;       
@@ -138,11 +136,11 @@ public class OrderItems implements Serializable {
         this.itemPriceRetail = itemPriceRetail;
     }
 
-    public int getItemDiscount() {
+    public double getItemDiscount() {
         return itemDiscount;
     }
 
-    public void setItemDiscount(int itemDiscount) {
+    public void setItemDiscount(double itemDiscount) {
         this.itemDiscount = itemDiscount;
     }
